+++ app/src/main/java/com/sopt/dive/network/model/signin/SignInResponse.kt
@@ -19,5 +19,5 @@ data class SignInResponseDataDto(
     @SerialName("message")
     val message: String,
     @SerialName("userId")
-    val userId: Int
+    val userId: Long
 )
